+++ core/util/dev/write.js
