+++ package.json
@@ -30,11 +30,5 @@
     "earliest",
     "ranges",
     "aliases"
-  ],
-  "standard": {
-    "globals": [
-      "describe",
-      "it"
-    ]
-  }
+  ]
 }
